enabled execution of audit-repo from any CWD
the script assumed you were in the check-for-secrets repo; now it can be
executed regardless of your current working directory.

also cleaned up the messaging a bit, such that it only says it's
checking for secrets once it actually is, such that it tells you the
full command it's running, etc.